requirements.txt
django==1.8.8
django-cors-headers
djangorestframework==3
